Cosmo Force: Moon spawn speed scales with attack speed (also applies to SoT, probably needs rebalance but we're mid rebalance anyway)

+++ Core/ModPlayers/AccessoryFieldsPlayer.cs
@@ -54,7 +54,7 @@ namespace FargowiltasSouls.Core.ModPlayers
 
         // new forces
         public int TerraProcCD;
-        public int CosmosMoonTimer;
+        public float CosmosMoonTimer;
         public int CosmosMoonCycle;
         public bool CosmosForce;
         public bool LifeForceActive;
